# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_doLogin_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 255
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  }
                },
                "required": ["identifier", "password"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Connexion (US-1.4) — renvoie les jetons ou un défi MFA",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/login/mfa": {
      "post": {
        "operationId": "AuthController_loginMfa[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challengeToken": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 512
                  },
                  "code": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 12
                  }
                },
                "required": ["challengeToken", "code"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Second facteur (TOTP, SMS ou code de récupération)",
        "tags": ["Authentification"]
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Déconnexion : révocation du refresh token et blacklist de l’access token",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me[0]_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mon compte : rôle, statut, état d’accès, MFA",
        "tags": ["Authentification"]
      }
    },
...
    "/api/v1/auth/sessions": {
      "get": {
        "operationId": "AuthController_sessions_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mes sessions actives (5 maximum)",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/sessions/{id}": {
      "delete": {
        "operationId": "AuthController_revokeSession_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Révoquer une de mes sessions",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/request-account": {
      "post": {
        "operationId": "AuthController_requestAccount[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "^[\\p{L}\\p{M}' .-]+$"
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "^[\\p{L}\\p{M}' .-]+$"
                  },
                  "email": {
                    "type": "string",
                    "maxLength": 255,
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "preferredChannel": {
                    "type": "string",
                    "enum": ["SMS", "EMAIL"]
                  },
                  "invitationCode": {
                    "type": "string",
                    "maxLength": 64
                  },
                  "tontineName": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "captchaToken": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4096
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "phone",
                  "preferredChannel",
                  "captchaToken"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Demande de compte (US-1.3) — réponse générique, captcha, 5 / heure / IP",
        "tags": ["Authentification"]
      }
    },
...
    "/api/v1/auth/activate": {
      "post": {
        "operationId": "AuthController_activate[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 512
                  },
                  "identifier": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 255
                  },
                  "otp": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["password"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Activation du compte par lien (48 h) ou OTP (15 min) et choix du mot de passe",
        "tags": ["Authentification"]
      }
    },
...
    "/api/v1/auth/activation/resend": {
      "post": {
        "operationId": "AuthController_resend_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 255
                  }
                },
                "required": ["identifier"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Renvoi du code d’activation (3 / heure)",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgot_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 255
                  }
                },
                "required": ["identifier"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mot de passe oublié (US-1.5) — toujours 200, sans fuite d’information",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_reset_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 512
                  },
                  "newPassword": {
                    "type": "string"
                  }
                },
                "required": ["token", "newPassword"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Nouveau mot de passe (jeton 1 h, usage unique, différent des 5 derniers)",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/mfa": {
      "get": {
        "operationId": "AuthController_mfaStatus_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "État de la double authentification",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/mfa/enable": {
      "post": {
        "operationId": "AuthController_mfaEnable[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["TOTP", "SMS"]
                  }
                },
                "required": ["type"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activer le MFA (US-1.6) : TOTP → QR code ; SMS → code envoyé",
        "tags": ["Authentification"]
      }
    },
...
    "/api/v1/auth/mfa/verify": {
      "post": {
        "operationId": "AuthController_mfaVerify_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": ["code"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirmer le MFA ; renvoie 10 codes de récupération (affichés une seule fois)",
        "tags": ["Authentification"]
      }
    },
    "/api/v1/auth/mfa/sms-code": {
      "post": {
        "operationId": "AuthController_mfaSmsCode[0]_v1",
        "parameters": [],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recevoir un code SMS (MFA SMS actif)",
        "tags": ["Authentification"]
      }
    },
...
    "/api/v1/auth/mfa/disable": {
      "post": {
        "operationId": "AuthController_mfaDisable[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "code": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 12
                  }
                },
                "required": ["password", "code"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Désactiver le MFA : mot de passe + code valide (alerte envoyée)",
        "tags": ["Authentification"]
      }
    },
...
    "/api/v1/auth/mfa/recovery-codes": {
      "post": {
        "operationId": "AuthController_regenerate_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 12
                  }
                },
                "required": ["code"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Régénérer les codes de récupération",
        "tags": ["Authentification"]
      }
    },
...
  },
  "info": {
    "title": "TontineMoney API",
    "description": "API REST versionnée du MVP TontineMoney (monolithe modulaire). Erreurs au format Problem Details (RFC 9457). Montants en chaînes décimales (unités majeures) + devise ISO 4217. En-tête Idempotency-Key obligatoire sur les POST financiers.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tm_rt"
      }
    },
...
  }
}